backend/pyproject.toml
[project]
name = "msal-demo-backend"
version = "0.1.0"
description = "FastAPI backend for MSAL multi-API demo"
authors = [ { name = "Your Name", email = "you@example.com" } ]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.0",
  "mangum==0.17.0",
  "python-dotenv==1.0.0",
  "msal==1.31.1",
  "pyjwt==2.10.1",
  "cryptography==44.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest==8.0.0",
  "httpx==0.27.0"
]

[project.scripts]
start = "app.main:run_dev"

[tool.uv]
# uv specific settings (placeholder; can add index configuration later)

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
